test(observability): add unit tests for ListTrajectoryOApi
Cover all branches:
- success with start_time provided
- success with start_time nil (uses benefit service)
- success with nil service response
- invalid request (nil, zero workspace_id, empty trace_ids, empty trace_id)
- permission error
- rate limited
- user missing when start_time nil
- service error

+++ backend/modules/observability/application/openapi.go
@@ -1164,7 +1164,7 @@ func (o *OpenAPIApplication) ListTrajectoryOApi(ctx context.Context, req *openap
 			SpaceID:      req.GetWorkspaceID(),
 		})
 		if err == nil && benefitRes != nil {
-			finalStartTime = time.Now().UnixMilli() - int64(benefitRes.StorageDuration)*24*60*60*1000
+			finalStartTime = time.Now().UnixMilli() - benefitRes.StorageDuration*24*60*60*1000
 		}
 		startTime = &finalStartTime
 	}
